requirements.txt
streamlit
torch
transformers
Pillow
python-dotenv
groq
